test: move eventually helper into src/client/test-utils
Importing it from logs.test.ts made bun register logs tests when
running dlq/client/flow-control test files individually.
