+++ src/client/dlq.test.ts
@@ -5,7 +5,7 @@
 import { sleep } from "bun";
 import { afterAll, beforeAll, describe, expect, test } from "bun:test";
 import { Client } from "./client";
-import { eventually } from "./logs.test";
+import { eventually } from "./test-utils";
 import { MOCK_QSTASH_SERVER_URL, mockQStashServer, expectToReject } from "./workflow/test-utils";
 import type { HttpClient } from "./http";
