+++ src/client/flow-control.test.ts
@@ -4,7 +4,7 @@
 import { describe, expect, test } from "bun:test";
 import { Client } from "./client";
 import { MOCK_QSTASH_SERVER_URL, mockQStashServer, expectToReject } from "./workflow/test-utils";
-import { eventually } from "./logs.test";
+import { eventually } from "./test-utils";
 
 describe("FlowControl empty id guard", () => {
   test("should not send request when get is called with an empty string", async () => {
